% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ear_process}
\alias{ear_process}
\title{Calculates Auditory Nerve Images}
\usage{
ear_process(wav, inNumOfChannels, inFirstFreq, inFreqDist, inSampleFrequency)
}
\arguments{
\item{wav}{a sampled mono signal}

\item{inNumOfChannels}{is the number of channels for decimation}

\item{inFirstFreq}{frequency of first channel (in critical band units)}

\item{inFreqDist}{the the frequency bandwidth of the "auditory filter"}

\item{inSampleFrequency}{is the sample frequency of the input signal}
}
\value{
Auditory Image
}
\description{
Calculates Auditory Nerve Images
}
