\name{OnsetPatternFilter}
\alias{OnsetPatternFilter}
\title{Filter \code{OnsetPattern} Output}
\description{This function takes an (multi-channel) onset pattern generated by \code{\link{OnsetPattern}}, and outputs a (one-dimensional) signal in which a non-zero value means an onset occurred (the value is a measurement of the likeliness for the onset). Currently, the following (simple) strategy is followed:
an onset is detected at a certain moment, if a certain fraction of channels have an onset within a specific period of time and, the moment falls at least a minimum period behind the last detected onset.}
\usage{OnsetPatternFilter(inOnsetPattern, inSampleFreq)}
\arguments{
  \item{inOnsetPattern}{a matrix of size  \emph{n} x \emph{m} where \emph{n} is the number of channels and \emph{m} is the length in samples.}
  \item{inSampleFreq}{the sample frequency of the incoming signal (in Hz).}}
\value{
\item{OnsetSignal}{a 1 dimension signal having a non-zero value at detected onset-times (the clearer the onset, the higher the value).}
  \item{OnsetSignalFreq}{sample frequency of the result (same as \code{inSampleFreq}).}}
\author{Marc Vidal (\code{R} version). Based on the original code from \pkg{IPEM} Toolbox.}
\examples{data(SchumannKurioseGeschichte)
s <- SchumannKurioseGeschichte
ANIs <- CalcANI(s, 22050)
OPs <- OnsetPattern(ANIs$AuditoryNerveImage, ANIs$ANIFreq)
OnsetPatternFilter(OPs$OnsetPattern, OPs$OnsetPatternFreq)}
