\name{SchumannKurioseGeschichte}
\alias{SchumannKurioseGeschichte}
\docType{data}
\title{Schumann, "Kuriose Geschichte"}
\description{This is a sampled signal (22050) of a \code{WAV} file containing the first bars of the piece "Kuriose Geschichte" (Curious Story) from R. Schumann's op.15.}
\usage{data("SchumannKurioseGeschichte")}
\format{the sampled signal in \code{numeric} class.}
\source{ISMPL - V. Horowitz (piano).}
\examples{data(SchumannKurioseGeschichte)}
\keyword{datasets}
